Move `escape-highlight` file to `utils` folder and fix imports
