+++ src/connectors/hits/connectHits.ts
@@ -1,5 +1,6 @@
-import escapeHits, { TAG_PLACEHOLDER } from '../../lib/escape-highlight';
 import {
+  escapeHits,
+  TAG_PLACEHOLDER,
   checkRendering,
   createDocumentationMessageGenerator,
   addAbsolutePosition,
