+++ src/lib/utils/concatHighlightedParts.ts
@@ -1,5 +1,5 @@
 import { HighlightedParts } from '../../types';
-import { TAG_REPLACEMENT } from '../escape-highlight';
+import { TAG_REPLACEMENT } from './escape-highlight';
 
 export default function concatHighlightedParts(parts: HighlightedParts[]) {
   const { highlightPreTag, highlightPostTag } = TAG_REPLACEMENT;
