+++ src/lib/utils/getHighlightedParts.ts
@@ -1,4 +1,4 @@
-import { TAG_REPLACEMENT } from '../../lib/escape-highlight';
+import { TAG_REPLACEMENT } from './escape-highlight';
 
 export default function getHighlightedParts(highlightedValue: string) {
   const { highlightPostTag, highlightPreTag } = TAG_REPLACEMENT;
